+++ Konigsberg.Tests/Euler/Problem5Tests.cs
@@ -1,5 +1,3 @@
-using System.Collections.Generic;
-
 using Konigsberg.Euler;
 
 namespace Konigsberg.Tests.Euler;
@@ -14,13 +12,5 @@ public sealed class Problem5Tests : EulerTests
         var problem = new Problem5(limit);
         var solution = problem.Solve();
         Assert.Equal(expected, solution);
-    }
-
-    [Fact]
-    public void PrimesUpThruTest()
-    {
-        var expectedPrimes = new List<int> { 2, 3, 5, 7, 11 };
-        var primes = Problem5.PrimesUpThru(11);
-        Assert.Equal(expectedPrimes, primes);
     }
 }
