+++ Konigsberg/Libraries/IntegerExtensions.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 
 namespace Konigsberg.Libraries;
 
@@ -22,5 +23,59 @@ public static class IntegerExtensions
         }
 
         return true;
+    }
+
+    /// <summary>
+    /// Returns a factorization dictionary with the prime as the key and the exponent as the value.
+    /// </summary>
+    /// <param name="n"></param>
+    /// <returns></returns>
+    public static Dictionary<int, int> PrimeFactorization(this int n)
+    {
+        if (n <= 1) throw new ArgumentException(nameof(n), "Must be 2 or greater.");
+
+        var factorization = new Dictionary<int, int>();
+
+        var primes = PrimesUpThru(n);
+        foreach (var p in primes)
+        {
+            var e = 0; // init exponent
+            while (n % p == 0)
+            {
+                n /= p;
+                e += 1;
+            }
+
+            factorization[p] = e;
+        }
+
+        return factorization;
+    }
+
+    internal static List<int> PrimesUpThru(int n)
+    {
+        var rootLimit = (int) Math.Floor(Math.Sqrt(n));
+        var primes = new List<int>();
+
+        for (var primeCandidate = 2; primeCandidate <= n; primeCandidate++)
+        {
+            var isPrime = true; // default
+            for (var j = 2; j <= rootLimit; j++)
+            {
+                if (primeCandidate % j == 0)
+                {
+                    foreach (var p in primes)
+                    {
+                        if (j % p == 0)
+                        {
+                            isPrime = false;
+                        }
+                    }
+                }
+            }
+            if (isPrime) primes.Add(primeCandidate);
+        }
+
+        return primes;
     }
 }
